requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
neo4j
pandas
numpy
python-dotenv
